:recycle: api: moved flashAlgorithmConfigurationRepository to shared in order to respect the context separation rules
