+++ api/src/certification/flash-certification/domain/usecases/index.js
@@ -5,6 +5,7 @@ import { fileURLToPath } from 'node:url';
 import * as challengeRepository from '../../../../shared/infrastructure/repositories/challenge-repository.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
+import * as sharedFlashAlgorithmConfigurationRepository from '../../../shared/infrastructure/repositories/flash-algorithm-configuration-repository.js';
 import * as flashAlgorithmConfigurationRepository from '../../infrastructure/repositories/flash-algorithm-configuration-repository.js';
 import * as flashAlgorithmService from '../services/algorithm-methods/flash.js';
 
@@ -13,11 +14,13 @@ import * as flashAlgorithmService from '../services/algorithm-methods/flash.js';
  *
  * @typedef {challengeRepository} ChallengeRepository
  * @typedef {flashAlgorithmConfigurationRepository} FlashAlgorithmConfigurationRepository
+ * @typedef {sharedFlashAlgorithmConfigurationRepository} SharedFlashAlgorithmConfigurationRepository
  * @typedef {flashAlgorithmService} FlashAlgorithmService
  */
 const dependencies = {
   challengeRepository,
   flashAlgorithmConfigurationRepository,
+  sharedFlashAlgorithmConfigurationRepository,
   flashAlgorithmService,
 };
