+++ api/tests/certification/flash-certification/unit/domain/usecases/get-active-flash-assessment-configuration_test.js
@@ -4,16 +4,15 @@ import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 describe('#getActiveFlashAssessmentConfiguration', function () {
   it('should return the last configuration', async function () {
     // given
-    const flashAlgorithmConfigurationRepository = {
-      getMostRecent: sinon.stub(),
+    const configuration = domainBuilder.buildFlashAlgorithmConfiguration();
+    const sharedFlashAlgorithmConfigurationRepository = {
+      getMostRecent: sinon.stub().resolves(configuration),
     };
 
-    const configuration = domainBuilder.buildFlashAlgorithmConfiguration();
-
-    flashAlgorithmConfigurationRepository.getMostRecent.resolves(configuration);
-
     // when
-    const activeConfiguration = await getActiveFlashAssessmentConfiguration({ flashAlgorithmConfigurationRepository });
+    const activeConfiguration = await getActiveFlashAssessmentConfiguration({
+      sharedFlashAlgorithmConfigurationRepository,
+    });
 
     // then
     expect(activeConfiguration).to.deep.equal(configuration);
