+++ api/tests/tooling/domain-builder/factory/build-flash-algorithm-configuration.js
@@ -1,5 +1,5 @@
-import { FlashAssessmentAlgorithmConfiguration } from '../../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithmConfiguration.js';
 import { FlashAssessmentSuccessRateHandler } from '../../../../src/certification/flash-certification/domain/models/FlashAssessmentSuccessRateHandler.js';
+import { FlashAssessmentAlgorithmConfiguration } from '../../../../src/certification/shared/domain/models/FlashAssessmentAlgorithmConfiguration.js';
 
 export const buildFlashAlgorithmConfiguration = ({
   warmUpLength,
